Move PendingLabel, RejectedLabel and ApprovedLabel to StatusLabel directory
