+++ src/components/StatusLabel/ApprovedLabel/index.tsx
